Add Login + Fix Delete

+++ src/project1/controller/DeleteKhachHangController.java
@@ -2,13 +2,17 @@ package project1.controller;
 
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
+import java.util.ArrayList;
 
 import javax.swing.JButton;
+import javax.swing.JDialog;
 import javax.swing.JOptionPane;
 import javax.swing.JTable;
 
 import project1.model.KhachHang;
 import project1.model.KhachHangDB;
+import project1.model.MuonXe;
+import project1.model.MuonXeDB;
 import project1.view.ButtonKhachHangView;
 import project1.view.MainUI;
 import project1.view.PanelQuanLyKhachHangView;
@@ -61,9 +65,18 @@ public class DeleteKhachHangController {
 							 JOptionPane.QUESTION_MESSAGE, null, null, null);
 					if(select == 0) {
 						String id = getId(index, 0);
-						khachHang = khachHangDB.getKhachHang(id);
-						khachHangDB.deleteKhachHang(khachHang);
-						tableKhachHangView.updateTable(khachHangDB.getAllKhachHang());
+						
+						if (!checkUser(id)) {
+							JOptionPane.showMessageDialog(new JDialog(), "Khách hàng đã hoặc đang trong quan hệ thuê trả \n " +
+																		  "Hãy xóa các thuê trả liên quan đến độc giả này");
+							return;
+						}
+						else {
+							khachHang = khachHangDB.getKhachHang(id);
+							khachHangDB.deleteKhachHang(khachHang);
+							tableKhachHangView.updateTable(khachHangDB.getAllKhachHang());
+						}
+						
 					}
 				}
 				else {
@@ -73,6 +86,12 @@ public class DeleteKhachHangController {
 		});
 	}
 	
-	
+	private boolean checkUser (String id) {
+		ArrayList<MuonXe> listRentXe = new MuonXeDB().getAllMuonXe();
+		for (int i = 0; i < listRentXe.size(); i++) {
+			if (listRentXe.get(i).getMaKH().equals(id)) return false;
+		}
+		return true;
+	}
 	
 }
